% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eo_to_centroids.R
\name{eo_to_centroids}
\alias{eo_to_centroids}
\title{Converts a shapefile with multipart element occurrence (EO) polygons into centroid coordinates for each polygon.}
\usage{
eo_to_centroids(eo_shp, sname = "scientificName")
}
\arguments{
\item{eo_shp}{Shapefile or path to shapefile containing EOs (usually downloaded from Biotics)}

\item{sname}{Species or community name column. Default = "scientificName"}
}
\value{
A data frame with sname and coordinates for each centroid
}
\description{
Converts a shapefile with multipart element occurrence (EO) polygons into centroid coordinates for each polygon.
}
