% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_eo_shp.R
\name{filter_eo_shp}
\alias{filter_eo_shp}
\title{Pre-filters element occurrences (EOs) based on specified attribute values}
\usage{
filter_eo_shp(
  eo_shp,
  remove = list(EO_Rank = c("X", "X?"), id_question = "Y"),
  sname = "scientificName"
)
}
\arguments{
\item{eo_shp}{Shapefile or path to shapefile containing EOs (usually
downloaded from Biotics)}

\item{remove}{Named list specifying columns and corresponding values to remove.
For example, \code{list(EO_Rank = c("X", "X?"), id_question = "Y")}.
Names must correspond to columns in \code{eo_shp}. Default is
\code{list(EO_Rank = c("X", "X?"), id_question = "Y")}.}

\item{sname}{Species or community name column. Default = "scientificName"}
}
\value{
An sf object with specified EOs removed
}
\description{
Removes EOs where specified attribute columns contain specified values.
Missing (NA) values are retained. By default, removes EOs with an
EO_Rank of X or X? and EOs flagged with a questionable ID.
}
