% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_eos.R
\name{tabulate_eos}
\alias{tabulate_eos}
\title{Determine number of pre-defined element occurrences (EOs) (unique EO_ID) and number of good EOs based on user-provided shapefile}
\usage{
tabulate_eos(
  eo_shp,
  EOID_col = "EOID",
  EO_rank_col = "EORANK",
  sname = "scientificName"
)
}
\arguments{
\item{eo_shp}{Shapefile or path to shapefile containing EOs (usually downloaded from Biotics)}

\item{EOID_col}{Required name containing unique identifier (e.g. EO_ID, ObjectID), for each pre-defined occurrence. Default = "EORANK"}

\item{EO_rank_col}{Optional name containing overall EO_rank (e.g. A, AB, D, etc). Default = "EORANK"}

\item{sname}{Species or community name column. Default = "scientificName"}
}
\value{
A data frame with number of EOs and number of good EOs organized by species or community name
}
\description{
Determine number of pre-defined element occurrences (EOs) (unique EO_ID) and number of good EOs based on user-provided shapefile
}
