% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkCAT_v3.R
\name{calc_threats}
\alias{calc_threats}
\title{Show the effects of different potential threat options on calculated ranks.}
\usage{
calc_threats(input_df, points_col = "Points")
}
\arguments{
\item{input_df}{A data frame containing the records to deduplicate {usually an output of runBulkCAT()}}

\item{points_col}{A string specifying the column name which contains rarity-based points.
Defaults to "Points".}
}
\value{
A data frame calculated SRanks and Points for different threat options.
}
\description{
Includes low, medium, and high threat options, calculating SRank_lowT, SRank_medT, SRank_highT
to assist with review of rarity-based ranks.
}
