% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_threats.R
\name{calc_threats}
\alias{calc_threats}
\title{Show the effects of different potential threat options on calculated ranks.}
\usage{
calc_threats(
  input_df,
  rarity_points_col = "PointsSpatial",
  trend_points_col = NULL
)
}
\arguments{
\item{input_df}{A data frame containing spatial rarity points by species {usually an output of runBulkCAT()}}

\item{rarity_points_col}{A string specifying the column name which contains rarity-based points.
Defaults to "PointsSpatial".}

\item{trends_col}{A string specifying the column name which contains trend-based points.
Defaults to NULL.}
}
\value{
A data frame calculated SRanks and Points for different threat options.
}
\description{
Includes low, medium, and high threat options, calculating SRank_lowT, SRank_medT, SRank_highT
to assist with review of rarity-based ranks.
}
