% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkCAT_v2.R
\name{run_bulkCAT}
\alias{run_bulkCAT}
\title{Run BulkCAT analysis}
\usage{
run_bulkCAT(
  input_df,
  sname = "acceptedName",
  lat = "decimalLatitude",
  lon = "decimalLongitude",
  eo_separation = 1000,
  grid_size = 2000
)
}
\arguments{
\item{input_df}{Input multispecies observation points dataframe}

\item{sname}{Species name column. Default = "acceptedName"}

\item{lat}{Latitude column name. Default = "decimalLatitude"}

\item{lon}{Longitude column name. Default = "decimalLongitude"}

\item{eo_separation}{Minimum separation distance (m) for unique EO clusters. Default = 1000m.}

\item{grid_size}{Side length (m) for AOO grid cells. Default = 2000m.}
}
\value{
Dataframe with calculated rarity metrics for each species/element.
}
\description{
This function completes conservation status assessments (rarity ranks) using NatureServe
methodology. BulkCAT requires a user-curated multi-species point observations dataset with lat/lon columns.
}
